ML/pyproject.toml
[project]
name = "mumhacks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.2.8",
    "faiss-cpu>=1.13.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.2.0",
    "langchain-groq>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
]
